split MailsF nested classes to separete classes, more unit-tests
