+++ mail/src/test/java/org/huberb/ee8sample/mail/DumbsterSendMessageIT.java
@@ -28,8 +28,8 @@ import javax.mail.Session;
 import javax.mail.Transport;
 import javax.mail.internet.InternetAddress;
 import javax.mail.internet.MimeMessage;
-import org.huberb.ee8sample.mail.MailsF.MimeMessageF;
-import org.huberb.ee8sample.mail.MailsF.SessionF;
+import org.huberb.ee8sample.mail.MimeMessageF;
+import org.huberb.ee8sample.mail.SessionF;
 import org.huberb.ee8sample.mail.Supports.ConsumerThrowingMessagingException;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.params.ParameterizedTest;
