+++ mail/src/test/java/org/huberb/ee8sample/mail/MailsFTest.java
@@ -22,10 +22,10 @@ import javax.mail.MessagingException;
 import javax.mail.Session;
 import javax.mail.internet.InternetAddress;
 import javax.mail.internet.MimeMessage;
-import org.huberb.ee8sample.mail.MailsF.InternetAddressBuilder;
-import org.huberb.ee8sample.mail.MailsF.InternetAddressF;
-import org.huberb.ee8sample.mail.MailsF.MimeMessageF;
-import org.huberb.ee8sample.mail.MailsF.SessionF;
+import org.huberb.ee8sample.mail.InternetAddressBuilder;
+import org.huberb.ee8sample.mail.InternetAddressF;
+import org.huberb.ee8sample.mail.MimeMessageF;
+import org.huberb.ee8sample.mail.SessionF;
 import org.huberb.ee8sample.mail.Supports.ConsumerThrowingMessagingException;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
