+++ mail/src/test/java/org/huberb/ee8sample/mail/MessageFormatBodyMergerTest.java
@@ -24,8 +24,8 @@ import javax.mail.MessagingException;
 import javax.mail.Session;
 import javax.mail.internet.MimeMessage;
 import org.huberb.ee8sample.mail.BodyTextMergers.MessageFormatBodyMerger;
-import org.huberb.ee8sample.mail.MailsF.MimeMessageF;
-import org.huberb.ee8sample.mail.MailsF.SessionF;
+import org.huberb.ee8sample.mail.MimeMessageF;
+import org.huberb.ee8sample.mail.SessionF;
 import org.huberb.ee8sample.mail.Supports.ConsumerThrowingMessagingException;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
