+++ mail/src/test/java/org/huberb/ee8sample/mail/StringFormatBodyMergerTest.java
@@ -25,8 +25,8 @@ import javax.mail.MessagingException;
 import javax.mail.Session;
 import javax.mail.internet.MimeMessage;
 import org.huberb.ee8sample.mail.BodyTextMergers.StringFormatBodyMerger;
-import org.huberb.ee8sample.mail.MailsF.MimeMessageF;
-import org.huberb.ee8sample.mail.MailsF.SessionF;
+import org.huberb.ee8sample.mail.MimeMessageF;
+import org.huberb.ee8sample.mail.SessionF;
 import org.huberb.ee8sample.mail.Supports.ConsumerThrowingMessagingException;
 import org.huberb.ee8sample.mail.Supports.MailRuntimeException;
 import static org.junit.jupiter.api.Assertions.assertEquals;
